6684 - tests & spliting helper into two files
