+++ shared/src/business/useCaseHelper/pendingItems/fetchPendingItems.js
@@ -1,7 +1,3 @@
-const { Case } = require('../../entities/cases/Case');
-const { DocketEntry } = require('../../entities/DocketEntry');
-const { omit, pick } = require('lodash');
-
 /**
  * fetchPendingItems
  *
@@ -11,12 +7,7 @@ const { omit, pick } = require('lodash');
  * @param {string} providers.docketNumber the optional docketNumber filter
  * @returns {Array} the pending items found
  */
-exports.fetchPendingItems = async ({
-  applicationContext,
-  docketNumber,
-  judge,
-  page,
-}) => {
+exports.fetchPendingItems = async ({ applicationContext, judge, page }) => {
   const source = [
     'associatedJudge',
     'caseCaption',
@@ -28,63 +19,18 @@ exports.fetchPendingItems = async ({
     'receivedAt',
   ];
 
-  // TODO: refactor, these are two major different paths
-  if (docketNumber) {
-    const caseResult = await applicationContext
-      .getPersistenceGateway()
-      .getCaseByDocketNumber({
-        applicationContext,
-        docketNumber,
-      });
-    const caseEntity = new Case(caseResult, { applicationContext });
-    let foundCases = [pick(caseEntity.validate().toRawObject(), source)];
-    let foundDocuments = [];
-
-    foundCases.forEach(foundCase => {
-      const foundCaseEntity = new Case(foundCase, {
-        applicationContext,
-      });
-
-      foundCaseEntity.docketEntries.forEach(document => {
-        if (document.pending && document.servedAt) {
-          foundDocuments.push({
-            ...omit(
-              new DocketEntry(
-                {
-                  ...document,
-                },
-                { applicationContext },
-              ).toRawObject(),
-              'entityName',
-            ),
-            associatedJudge: foundCaseEntity.associatedJudge,
-            caseCaption: foundCaseEntity.caseCaption,
-            docketNumber: foundCaseEntity.docketNumber,
-            docketNumberSuffix: foundCaseEntity.docketNumberSuffix,
-            status: foundCaseEntity.status,
-          });
-        }
-      });
+  const pendingItemResults = await applicationContext
+    .getPersistenceGateway()
+    .fetchPendingItems({
+      applicationContext,
+      judge,
+      page,
+      source,
     });
-
-    return {
-      foundDocuments,
-      total: foundDocuments.length,
-    };
-  } else {
-    const pendingItemResults = await applicationContext
-      .getPersistenceGateway()
-      .fetchPendingItems({
-        applicationContext,
-        judge,
-        page,
-        source,
-      });
-    const foundDocuments = pendingItemResults.results;
-    const documentsTotal = pendingItemResults.total;
-    return {
-      foundDocuments,
-      total: documentsTotal,
-    };
-  }
+  const foundDocuments = pendingItemResults.results;
+  const documentsTotal = pendingItemResults.total;
+  return {
+    foundDocuments,
+    total: documentsTotal,
+  };
 };
