+++ shared/src/business/useCases/pendingItems/generatePrintablePendingReportInteractor.test.js
@@ -16,23 +16,25 @@ describe('generatePrintablePendingReportInteractor', () => {
       upload: jest.fn((params, callback) => callback()),
     });
 
-    applicationContext.getUseCaseHelpers().fetchPendingItems.mockReturnValue([
-      {
-        associatedJudge: 'Judge Colvin',
-        caseCaption: 'Test Caption, Petitioner',
-        docketNumber: '123-45',
-        documentTitle: 'Test Document Title',
-        receivedAt: '2020-01-01T12:00:00.000Z',
-      },
-      {
-        associatedJudge: 'Judge Buch',
-        caseCaption: 'Test Caption Two, Petitioner(s)',
-        docketNumber: '234-56',
-        docketNumberSuffix: DOCKET_NUMBER_SUFFIXES.SMALL,
-        documentType: 'Test Document Type',
-        receivedAt: '2020-02-02T12:00:00.000Z',
-      },
-    ]);
+    applicationContext.getUseCaseHelpers().fetchPendingItems.mockReturnValue({
+      foundDocuments: [
+        {
+          associatedJudge: 'Judge Colvin',
+          caseCaption: 'Test Caption, Petitioner',
+          docketNumber: '123-45',
+          documentTitle: 'Test Document Title',
+          receivedAt: '2020-01-01T12:00:00.000Z',
+        },
+        {
+          associatedJudge: 'Judge Buch',
+          caseCaption: 'Test Caption Two, Petitioner(s)',
+          docketNumber: '234-56',
+          docketNumberSuffix: DOCKET_NUMBER_SUFFIXES.SMALL,
+          documentType: 'Test Document Type',
+          receivedAt: '2020-02-02T12:00:00.000Z',
+        },
+      ],
+    });
 
     applicationContext
       .getPersistenceGateway()
